+++ khorosjx/admin.py
@@ -6,13 +6,13 @@
 :Example:        Coming Soon
 :Created By:     Jeff Shurtliff
 :Last Modified:  Jeff Shurtliff
-:Modified Date:  21 Sep 2021
+:Modified Date:  22 Sep 2021
 """
 
 from . import core
 
 # Define global variables
-base_url, api_credentials = None, None
+base_url, api_credentials = '', None
 
 
 # Define function to verify the connection in the core module